+++ feature/feature-home/src/main/java/com/comit/feature_home/navigation/HomeNavigation.kt
@@ -11,7 +11,7 @@ import androidx.navigation.navigation
 import com.comit.feature_home.screen.HomeScreen
 import com.comit.feature_home.screen.SalaryWebViewScreen
 import com.comit.feature_home.screen.alarm.AlarmScreen
-import com.comit.feature_home.screen.closeday.WriteClosedDayScreen
+import com.comit.feature_home.vm.WriteClosedDayScreen
 import com.comit.feature_home.screen.schedule.ShowScheduleScreen
 import com.comit.feature_home.screen.schedule.WriteScheduleScreen
 import com.comit.navigator.SimTongScreen
