+++ feature/feature-home/src/main/java/com/comit/feature_home/vm/WriteClosedDayScreen.kt
@@ -1,10 +1,10 @@
 @file:OptIn(
-    ExperimentalMaterialApi::class, ExperimentalMaterialApi::class,
-    InternalCoroutinesApi::class, ExperimentalMaterialApi::class
+    ExperimentalMaterialApi::class,
+    InternalCoroutinesApi::class,
 )
 @file:Suppress("OPT_IN_IS_NOT_ENABLED")
 
-package com.comit.feature_home.screen.closeday
+package com.comit.feature_home.vm
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -62,7 +62,6 @@ import com.comit.feature_home.calendar.SimTongCalendar
 import com.comit.feature_home.calendar.SimTongCalendarStatus
 import com.comit.feature_home.mvi.CloseDaySideEffect
 import com.comit.feature_home.string
-import com.comit.feature_home.vm.CloseDayViewModel
 import com.example.feature_home.R
 import kotlinx.coroutines.InternalCoroutinesApi
 import kotlinx.coroutines.launch
